gcd/requirements.txt
joblib
google-cloud-storage
twilio
scikit-learn
flask
pandas
flask_cors